pyproject.toml
[project]
name = "deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.0.0",
    "deepagents>=0.4.5",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langchain-mcp-adapters>=0.1.9",
    "langsmith>=0.7.13",
    "langchain-tavily>=0.2.17",
    "langgraph-cli[inmem]>=0.4.14",
    "markdownify>=1.2.2",
    "openai>=2.26.0",
    "PyYAML>=6.0.2",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "tavily-python>=0.7.22",
]

[[tool.uv.index]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
